ZBEARs_Weight_Smacher_v_01_08.py
A good loop phrase can change the world.
A rock sound with some meaning can change the world.
=> zbear on internet


Use of this tool:

- You can see the guide according to your ability or tendency.


-  Description Viewpoint A : Public service performance style

The mesh part which is confused by the various cases and operates is registered as the new vertex group and all vertices in the registered vertex group are liberated from all interference by the unknown several bones. This new vertex group's weights will be reset by the weight value given your inputted float value for an additional specified bone. This simplifies the complex situation affected by various bones. This simplifies the indiscriminate weighting of a vertices part caused by automatic weight paint, and helps to keep the work going by making the weighting situation of irregular mesh simple again without backing down from the middle point of the work.

UI No. 0: There must be an active object in selection, and its name must be on UI.
UI No. 1: Collect the Vertices that have been interfered with and lost their way for all of armature's Bones, designate them as the new Vertex Group and prepare new name for this Group.
UI No. 2: Enter the name of the new vertex group above.
UI No. 3: Decide one bone to be attached on your New Vertex Group (if the weighting states of this vertex group are already nice, that vertices in that group's will return to the position of their non-weighted original local location).
UI No. 4: Enter the name of the bone you picked from above.
UI No. 5: Enter the weight value to decide how much the Vertex Group will be affected by the Bone entered by you

UI No. 6: Check all parts and click this button if there is no logical flaw.


-  Description Viewpoint B :  Casual style


This tool can be used at more proper situation through the way of resetting partial weight to 0 and giving weights only to one of the most relevant bones. It would be better than keep your working state in the following situations :

Specially, in case of the mesh which is skinned with all kinds of the automatic weighting techniques, the weighting result is partly very complicatedly twisted and it shows the undesirable deforming shapes in the pose change

Use as an assistant tool of weighting solution when it's  needed to correct accidental weighting twisting by various automatic weighting techniques

When the weights of the vertices blocked from the view were painted weirdly by multiple bones, and even this situation was accumulated, and UNDO function is no more helpful to recover, or you feel like the one more UNDO can shutting down the Blender

When using weight painting mode and even if you still clicking Ctrl + LMB on each bones for more than 40 minutes but there is no better

Even doing one more action with Ctrl + Bone clicking is disgustingly tired and mentally painful


Hello, I'm ZBEAR. The Hard Rock Guitarist and 3D Music Video maker using Blender.

First, before I say something, I am sorry about my hard rock promotion, which is decided 
as much as to make and distribute functions for people free of charge.
I am very sorry that my urgent problem, the global hard rock inflation problem, 
has led to the inclusion of advertisements for my artist's name 'ZBEAR' and its activity, 
hard rock guitar solo albums, and if the world could have prevented me from becoming such a 
hungry guitarist by blind capitalism(I support capitalism system with not-blinded options), 
all of these advertisements in everyside of this add-on could have been removed. I want you to understand.

This Code has simple routine but I thought the idea for making this feature would be worthwhile in that no one tried to implement this convenient function.
So, despite having to burn out the guitar strings as a guitarist, I decided to make this add-on while making music videos using Great Tool called Blender.

In many cases, someone may not be aware of why this add-on's core vision is currently needed for them.
However, if it is used properly at the proper situation, this idea can save your non-productive time by selectively initializing the values ​​of the complexly twisted weighting table and re-painting the weights for the mesh part that comes back into view.
Check out the description on the product page and packed readme files for more details.

'And if possible, if someday the blender team decide to include this partial weighting management idea of ​​this addon, I would be very grateful if that function is added with marking my active name 'ZBEAR'.'